api2: moved genesis generation cli->api
